refactor(cli): consolidate activity logger registration and improve storage integration
